Move break flag to SentakkiHitObject
