+++ osu.Game.Rulesets.Sentakki/Objects/Drawables/DrawableSentakkiLanedHitObject.cs
@@ -1,11 +1,5 @@
 ﻿using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Game.Rulesets.Objects;
-using osu.Game.Rulesets.Objects.Drawables;
-using osu.Game.Rulesets.Scoring;
 using osu.Game.Rulesets.Sentakki.UI;
-using osu.Game.Skinning;
 
 namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
 {
@@ -16,9 +10,6 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
         protected override float SamplePlaybackPosition =>
             (SentakkiExtensions.GetPositionAlongLane(SentakkiPlayfield.INTERSECTDISTANCE, HitObject.Lane).X / (SentakkiPlayfield.INTERSECTDISTANCE * 2)) + .5f;
 
-        private PausableSkinnableSound breakSample = null!;
-
-        private Container<DrawableScorePaddingObject> scorePaddingObjects = null!;
 
         public DrawableSentakkiLanedHitObject(SentakkiLanedHitObject? hitObject)
             : base(hitObject)
@@ -28,81 +19,8 @@ namespace osu.Game.Rulesets.Sentakki.Objects.Drawables
         [BackgroundDependencyLoader]
         private void load()
         {
-            AddRangeInternal(new Drawable[]
-            {
-                scorePaddingObjects = new Container<DrawableScorePaddingObject>(),
-                breakSample = new PausableSkinnableSound(),
-            });
-
             if (DrawableSentakkiRuleset is not null)
                 AnimationDuration.BindTo(DrawableSentakkiRuleset?.AdjustedAnimDuration);
-        }
-
-        protected override void LoadSamples()
-        {
-            base.LoadSamples();
-
-            LoadBreakSamples();
-        }
-
-        public void LoadBreakSamples()
-        {
-            breakSample.Samples = HitObject.CreateBreakSample();
-        }
-
-        public override void PlaySamples()
-        {
-            base.PlaySamples();
-
-            breakSample.Balance.Value = CalculateSamplePlaybackBalance(SamplePlaybackPosition);
-            breakSample.Play();
-        }
-
-        protected override DrawableHitObject CreateNestedHitObject(HitObject hitObject)
-        {
-            switch (hitObject)
-            {
-                case ScorePaddingObject p:
-                    return new DrawableScorePaddingObject(p);
-            }
-
-            return base.CreateNestedHitObject(hitObject);
-        }
-
-        protected override void AddNestedHitObject(DrawableHitObject hitObject)
-        {
-            switch (hitObject)
-            {
-                case DrawableScorePaddingObject p:
-                    scorePaddingObjects.Add(p);
-                    break;
-
-
-                default:
-                    base.AddNestedHitObject(hitObject);
-                    break;
-            }
-        }
-
-        protected override void ClearNestedHitObjects()
-        {
-            base.ClearNestedHitObjects();
-            scorePaddingObjects.Clear(false);
-        }
-
-        protected override void OnFree()
-        {
-            base.OnFree();
-            breakSample.ClearSamples();
-        }
-
-        protected new void ApplyResult(HitResult hitResult)
-        {
-            // Also give Break note score padding a judgement
-            for (int i = 0; i < scorePaddingObjects.Count; ++i)
-                scorePaddingObjects[^(i + 1)].ApplyResult(hitResult);
-
-            base.ApplyResult(hitResult);
         }
     }
 }
