+++ osu.Game.Rulesets.Sentakki/Objects/SentakkiHitObject.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using Newtonsoft.Json;
 using osu.Framework.Bindables;
 using osu.Framework.Extensions.Color4Extensions;
@@ -16,6 +17,20 @@ namespace osu.Game.Rulesets.Sentakki.Objects
 {
     public abstract class SentakkiHitObject : HitObject
     {
+        public int ScoreWeighting => Break ? 5 : BaseScoreWeighting;
+
+        protected virtual int BaseScoreWeighting => 1;
+
+        public readonly BindableBool BreakBindable = new BindableBool();
+
+        protected virtual bool PlaysBreakSample => true;
+
+        public bool Break
+        {
+            get => BreakBindable.Value;
+            set => BreakBindable.Value = value;
+        }
+
         protected SentakkiHitObject()
         {
             // We initialize the note colour to the default value first for test scenes
@@ -47,6 +62,49 @@ namespace osu.Game.Rulesets.Sentakki.Objects
         public virtual Color4 DefaultNoteColour => Color4Extensions.FromHex("FF0064");
 
         protected override HitWindows CreateHitWindows() => new SentakkiTapHitWindows();
+
+        protected override void CreateNestedHitObjects(CancellationToken cancellationToken)
+        {
+            base.CreateNestedHitObjects(cancellationToken);
+
+            for (int i = 1; i < ScoreWeighting; ++i)
+                AddNested(new ScorePaddingObject { StartTime = this.GetEndTime() });
+        }
+
+        public HitSampleInfo[] CreateBreakSample()
+        {
+            if (!PlaysBreakSample || !Break)
+                return Array.Empty<HitSampleInfo>();
+
+            return new[]
+            {
+                new BreakSample( CreateHitSampleInfo())
+            };
+        }
+
+        public override IList<HitSampleInfo> AuxiliarySamples =>
+            (Break && PlaysBreakSample) ? [new BreakSample(CreateHitSampleInfo())] : [];
+
+        public class BreakSample : HitSampleInfo
+        {
+            public override IEnumerable<string> LookupNames
+            {
+                get
+                {
+                    foreach (string name in base.LookupNames)
+                        yield return name;
+
+                    foreach (string name in base.LookupNames)
+                        yield return name.Replace("-max", string.Empty);
+                }
+            }
+
+            public BreakSample(HitSampleInfo sampleInfo)
+                : base("spinnerbonus-max", sampleInfo.Bank, sampleInfo.Suffix, sampleInfo.Volume)
+
+            {
+            }
+        }
 
         // This special hitsample is used for Sentakki specific samples, with doesn't have bank specific variants
         public class SentakkiHitSampleInfo : HitSampleInfo, IEquatable<SentakkiHitSampleInfo>
